requirements.txt
aiofiles==23.1.0
anyio==3.7.0
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
orjson==3.9.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
sniffio==1.3.0
tqdm==4.65.0
twitter-api-client @ git+https://github.com/yibudak/twitter-api-client.git@068519344662594036b0c20fadaefaeb4bed9349
uvloop==0.17.0
websockets==11.0.3
Werkzeug==2.3.6
